requirements.txt
appdirs==1.4.4
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
Django==3.2.7
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-js-asset==1.2.2
django-rest-swagger==2.2.0
django-rosetta==0.9.7
djangorestframework==3.12.4
drf-yasg==1.20.0
fake-useragent==0.1.11
idna==3.2
importlib-metadata==4.8.1
inflection==0.5.1
instagramy==4.4
itypes==1.2.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
parse==1.19.0
Pillow==8.3.2
polib==1.1.1
pyee==8.2.2
pyparsing==2.4.7
pyppeteer==0.2.6
pyquery==1.4.3
pytz==2021.1
requests==2.26.0
requests-html==0.10.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tqdm==4.62.2
uritemplate==3.0.1
urllib3==1.26.6
w3lib==1.22.0
websockets==9.1
zipp==3.5.0
